+++ lib/roadie/document.rb
@@ -16,7 +16,7 @@ module Roadie
   # @attr [#call] before_transformation Callback to call just before {#transform}ation begins. Will be called with the parsed DOM tree and the {Document} instance.
   # @attr [#call] after_transformation Callback to call just before {#transform}ation is completed. Will be called with the current DOM tree and the {Document} instance.
   class Document
-    attr_reader :html, :asset_providers
+    attr_reader :html, :asset_providers, :external_asset_providers
 
     # URL options. If none are given no URL rewriting will take place.
     # @see UrlGenerator#initialize
@@ -24,10 +24,15 @@ module Roadie
 
     attr_accessor :before_transformation, :after_transformation
 
+    # Should CSS that cannot be inlined be kept in a new `<style>` element in `<head>`?
+    attr_accessor :keep_uninlinable_css
+
     # @param [String] html the input HTML
     def initialize(html)
+      @keep_uninlinable_css = true
       @html = html
       @asset_providers = ProviderList.wrap(FilesystemProvider.new)
+      @external_asset_providers = ProviderList.empty
       @css = ""
     end
 
@@ -66,9 +71,14 @@ module Roadie
       dom.dup.to_html
     end
 
-    # Assign new asset providers. The supplied list will be wrapped in a {ProviderList} using {ProviderList.wrap}.
+    # Assign new normal asset providers. The supplied list will be wrapped in a {ProviderList} using {ProviderList.wrap}.
     def asset_providers=(list)
       @asset_providers = ProviderList.wrap(list)
+    end
+
+    # Assign new external asset providers. The supplied list will be wrapped in a {ProviderList} using {ProviderList.wrap}.
+    def external_asset_providers=(list)
+      @external_asset_providers = ProviderList.wrap(list)
     end
 
     private
@@ -81,8 +91,8 @@ module Roadie
     end
 
     def inline(dom)
-      dom_stylesheets = AssetScanner.new(dom, asset_providers).extract_css
-      Inliner.new(dom_stylesheets + [stylesheet]).inline(dom)
+      dom_stylesheets = AssetScanner.new(dom, asset_providers, external_asset_providers).extract_css
+      Inliner.new(dom_stylesheets + [stylesheet], dom).inline(keep_uninlinable_css)
     end
 
     def rewrite_urls(dom)