+++ lib/roadie/filesystem_provider.rb
@@ -8,7 +8,6 @@ module Roadie
     # the base path.
     class InsecurePathError < Error; end
 
-    include AssetProvider
     attr_reader :path
 
     def initialize(path = Dir.pwd)
@@ -34,6 +33,9 @@ module Roadie
         raise CssNotFound.new(basename, %{#{file_path} does not exist. (Original name was "#{name}")}, self)
       end
     end
+
+    def to_s() inspect end
+    def inspect() "#<#{self.class} #@path>" end
 
     private
     def build_file_path(name)