+++ lib/roadie/inliner.rb
@@ -16,64 +16,94 @@ module Roadie
   #   <a href="/" style="color:red"></a>
   class Inliner
     # @param [Array<Stylesheet>] stylesheets the stylesheets to use in the inlining
-    def initialize(stylesheets)
+    # @param [Nokogiri::HTML::Document] dom
+    def initialize(stylesheets, dom)
       @stylesheets = stylesheets
+      @dom = dom
     end
 
     # Start the inlining, mutating the DOM tree.
     #
-    # @param [Nokogiri::HTML::Document] dom
+    # @param [true, false] keep_extra_blocks
     # @return [nil]
-    def inline(dom)
-      apply style_map(dom)
+    def inline(keep_extra_blocks = true)
+      style_map, extra_blocks = consume_stylesheets
+
+      apply_style_map(style_map)
+      add_styles_to_head(extra_blocks) if keep_extra_blocks
+
       nil
     end
 
+    protected
+    attr_reader :stylesheets, :dom
+
     private
-    attr_reader :stylesheets
-
-    def apply(style_map)
-      style_map.each_element do |element, builder|
-        apply_element_style element, builder
-      end
-    end
-
-    def style_map(dom)
+    def consume_stylesheets
       style_map = StyleMap.new
+      extra_blocks = []
 
-      each_inlinable_block do |stylesheet, selector, properties|
-        elements = elements_matching_selector(stylesheet, selector, dom)
-        style_map.add elements, properties
-      end
-
-      style_map
-    end
-
-    def each_inlinable_block
-      stylesheets.each do |stylesheet|
-        stylesheet.each_inlinable_block do |selector, properties|
-          yield stylesheet, selector, properties
+      each_style_block do |stylesheet, block|
+        if (elements = selector_elements(stylesheet, block))
+          style_map.add elements, block.properties
+        else
+          extra_blocks << block
         end
       end
+
+      [style_map, extra_blocks]
+    end
+
+    def each_style_block
+      stylesheets.each do |stylesheet|
+        stylesheet.blocks.each do |block|
+          yield stylesheet, block
+        end
+      end
+    end
+
+    def selector_elements(stylesheet, block)
+      block.inlinable? && elements_matching_selector(stylesheet, block.selector)
+    end
+
+    def apply_style_map(style_map)
+      style_map.each_element { |element, builder| apply_element_style(element, builder) }
     end
 
     def apply_element_style(element, builder)
       element["style"] = [builder.attribute_string, element["style"]].compact.join(";")
     end
 
-    def elements_matching_selector(stylesheet, selector, dom)
+    def elements_matching_selector(stylesheet, selector)
       dom.css(selector.to_s)
     # There's no way to get a list of supported pseudo selectors, so we're left
     # with having to rescue errors.
     # Pseudo selectors that are known to be bad are skipped automatically but
     # this will catch the rest.
     rescue Nokogiri::XML::XPath::SyntaxError, Nokogiri::CSS::SyntaxError => error
-      warn "Roadie cannot use #{selector.inspect} (from \"#{stylesheet.name}\" stylesheet) when inlining stylesheets"
-      []
+      Utils.warn "Cannot inline #{selector.inspect} from \"#{stylesheet.name}\" stylesheet. If this is valid CSS, please report a bug."
+      nil
     rescue => error
-      warn "Roadie got error when looking for #{selector.inspect} (from \"#{stylesheet.name}\" stylesheet): #{error}"
+      Utils.warn "Got error when looking for #{selector.inspect} (from \"#{stylesheet.name}\" stylesheet): #{error}"
       raise unless error.message.include?('XPath')
-      []
+      nil
+    end
+
+    def add_styles_to_head(blocks)
+      unless blocks.empty?
+        create_style_element(blocks, find_head)
+      end
+    end
+
+    def find_head
+      dom.at_xpath('html/head')
+    end
+
+    def create_style_element(style_blocks, head)
+      return unless head
+      element = Nokogiri::XML::Node.new("style", head.document)
+      element.content = style_blocks.join("\n")
+      head.add_child(element)
     end
 
     # @api private
@@ -94,8 +124,8 @@ module Roadie
         end
       end
 
-      def each_element
-        @map.each_pair { |element, builder| yield element, builder }
+      def each_element(&block)
+        @map.each_pair(&block)
       end
     end
   end