+++ lib/roadie/rspec.rb
@@ -1,1 +1,2 @@
 require 'roadie/rspec/asset_provider'
+require 'roadie/rspec/cache_store'